fix(#139): address PR D review — extract shared deriveMode, add mount fetch, clean up save handlers
Review round 1 fixes:
- Extract deriveCredentialSetMode + ENVIRONMENT to shared credentialSetMode.ts
  (eliminates duplicate logic in SettingsPage and SetupPage)
- Remove dead ternaries (const keyValue = mode === "repair" ? apiKey : apiKey)
- Flatten handleCreate/handleBrokerSubmit: save both keys first, then show
  phrase modal — eliminates confusing nested early-return structure
- Add useEffect mount fetch in SetupPage: call refreshCredentials() when
  step transitions to "broker" so partial-save state from a prior session
  is detected on entry
- Update SetupPage test mock queue: mount fetch consumes first
  mockResolvedValueOnce, post-error refresh consumes second
